requirements/production.txt
# Dependencies related to deploying frame-server

# WSGI Handler
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
django-storages==1.6.5

# Caching
hiredis==0.2.0
django-redis-cache==1.7.1

# Django-secure
django-secure==1.0.1

# App Monitoring
newrelic==2.98.0.81
